Fix a warning with USE_RUBY_DEBUG_LOG=1 on macOS
```
compiling ../thread.c
In file included from ../thread.c:263:
In file included from ../thread_pthread.c:2870:
../thread_pthread_mn.c:777:43: warning: format specifies type 'unsigned long' but the argument has type 'rb_hrtime_t' (aka 'unsigned long long') [-Wformat]
		RUBY_DEBUG_LOG("abs:%lu", abs);
				    ~~~   ^~~
				    %llu
../vm_debug.h:110:74: note: expanded from macro 'RUBY_DEBUG_LOG'
	ruby_debug_log(__FILE__, __LINE__, RUBY_FUNCTION_NAME_STRING, "" __VA_ARGS__); \
									 ^~~~~~~~~~~
1 warning generated.
```

+++ thread_pthread_mn.c
@@ -774,7 +774,7 @@ timer_thread_register_waiting(rb_thread_t *th, int fd, enum thread_sched_waiting
                 ccan_list_add_tail(&timer_th.waiting, &th->sched.waiting_reason.node);
             }
             else {
-                RUBY_DEBUG_LOG("abs:%lu", abs);
+                RUBY_DEBUG_LOG("abs:%lu", (unsigned long)abs);
                 VM_ASSERT(flags & thread_sched_waiting_timeout);
 
                 // insert th to sorted list (TODO: O(n))
